Flytter kode til egen Filopplaster2.tsx fil
